Add basic function for search text modification
See
https://github.com/ankitects/anki/blob/91238211318976f63e66724b16bee12985ae3081/qt/tools/genhooks_gui.py#L497
for a short explanation of this hook and why the
check whether context.ids has already been set is
recommended.
As a search text consisting of
several words is a good sign of a complex query we
don't want to mess around with and corrupt,
Advocator will only prefix individual words with
`nc:`.