Allow specifying deck name in addition to note type
To make the usage of Advocator easier, users are
now able to specify the name of the deck as well.
This way, a user does not have to create an extra
note type for Advocator to function properly, but
is free to use any note type for adding cards to
the previously specified deck, whithout having
to worry about vowels not getting replaced for
other cards using another note type.

Closes #10